+++ src/Symfony/Component/Form/Extension/Validator/ViolationMapper/ViolationMapper.php
@@ -88,8 +88,8 @@ class ViolationMapper implements ViolationMapperInterface
         }
 
         // This case happens if an error happened in the data under a
-        // virtual form that does not match any of the children of
-        // the virtual form.
+        // form inheriting its parent data that does not match any of the
+        // children of that form.
         if (null !== $violationPath && !$match) {
             // If we could not map the error to anything more specific
             // than the root element, map it to the innermost directly
@@ -162,7 +162,7 @@ class ViolationMapper implements ViolationMapperInterface
             }
         }
 
-        // Ignore virtual forms when iterating the children
+        // Skip forms inheriting their parent data when iterating the children
         $childIterator = new \RecursiveIteratorIterator(
             new VirtualFormAwareIterator($form->all())
         );
@@ -253,8 +253,8 @@ class ViolationMapper implements ViolationMapperInterface
             // Process child form
             $scope = $scope->get($it->current());
 
-            if ($scope->getConfig()->getVirtual()) {
-                // Form is virtual
+            if ($scope->getConfig()->getInheritData()) {
+                // Form inherits its parent data
                 // Cut the piece out of the property path and proceed
                 $propertyPathBuilder->remove($i);
             } elseif (!$scope->getConfig()->getMapped()) {